* v3: (70 commits)
  Renamed axisMinValue/axisMaxValue to axisMinimum/axisMaximum
  Set default barWidth to match old default behaviour
  Updated tests
  Added more samples
  Added helper to generate a fill formatter with a closure (block)
  Fixed xAxis min/max calculation typo
  These should be internal
  More improvements to examples
  Polished demos
  Some minor bug fixes for formatters
  Corrected formula for 29-feb :-)
  Avoid truncating using Int(...), it overflows
  Renamed formatter interfaces
  This workaround is not required in the newer compilers
  Avoid x values in Pie's calcMinMax
  Breaking change: `valueFormatter` now accepts an IValueFormatter
  Cache XBounds object
  Removed the extra offset that messed up legend location
  Minor refactoring
  Fixed combined chart highlight with bar data
  ...

+++ ChartsDemo/Classes/RealmBase/Score.h
@@ -11,11 +11,11 @@
 @interface Score : RLMObject
 
 @property (nonatomic, assign) float totalScore;
-@property (nonatomic, assign) NSInteger scoreNr;
+@property (nonatomic, assign) double scoreNr;
 @property (nonatomic, strong) NSString *playerName;
 
 - (id)initWithTotalScore:(float)totalScore
-                 scoreNr:(NSInteger)scoreNr
+                 scoreNr:(double)scoreNr
               playerName:(NSString *)playerName;
 
 @end
